solar-vue-web-backend/requirements.txt
Flask==2.1.1
Flask-Limiter==2.2.0
Flask-Login==0.6.0
glob2==0.7
gunicorn==20.1.0
numpy==1.22.3
opencv-python-headless==4.5.5.64
pandas==1.4.1
PyJWT==2.3.0
pymongo==4.0.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.0
pdf2image==1.16.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0